Decrease performance impact of row systems, improve overall behavior
Row systems initially had to either exactly match a set of components that was added to an entity, or, in the case the row system on matched one compoent, had to match one of the components added to an entity.

This led to cases where it was hard to determine why a row system wasn't invoked (a component was missing, or too many were added). It also reduced how generally applicable row systems were, as applications needed to have full control over which components get added, removed or set.

The old design also had a much larger performance impact, where if a row system wasn't exactly matched, reflecs had to check for every added component, whether it was matched by the row system. This led to many lookups which slowed down adding/removing components.

In the new design, a single O(1) lookup will determine whether a row system needs to be executed for an add/remove/set operation. Furthermore, row systems are matched in advance to *families*. Because this does not happen in the critical path, reflecs can now do proper intersection matching on the set of components of the system and the set of components that are added.

The only operation in the new design that is more expensive, is creating a new family. This is a one-time operation per family, and if an application desires, can be forced to take place before the main loop.