Always stage tables, even when iterating in main thread
The reason for this is that notifications for new tables could get lost when created while iterating. This modification ensures tables are always created while either merging or not iterating, which simplifies notifications.

Notifications cannot be sent while iterating, as it modifies system metadata which can interfere with the logic that is running systems.