Fix issue with multiple threads & keeping track of absolute offset
Systems receive an 'offset' and 'frame_offset'. The offset is the row in the archetype from which the system should start processing. Typically this is 0, but in some cases (when using the offset,limit arguments, or when distributing load across worker threads) this number can be non-zero.

A system is called once for every archetype. Sometimes a system needs to know not just the offset relative to the current archetype, but the offset relative to all entities that match its signature. This is the frame_offset. This commit fixes issues with the frame offset (previously called index_offset), and changes the API to simplify system iterations, while also making it easier to work with frame_offset.