Delay notifying systems of tables until after column eval
During a merge tables can be created, which have to be matched against systems. In order for systems to be matched to tables with prefabs, the table must be registered in the prefab index. The prefab index is updated when columns are evaluated during a merge, while systems were notified before that, which caused this issue.