Fix issue with creating prefab instances in progress
When a prefab was instantiated for the first time while in progress, the prefab index was not yet updated, so flecs did not correctly detect that the type required a prefab.

To address this, a new flag has been added to detect when a table  has just been added. When this flag is set, flecs will look in the data stage to detect whether the entity has a prefab.