Implement algorithm to ensure entities are unique in a type
Flags make it more difficult to merge types, as the index of an entity in a type is not just determined by its id, but also its flags. To keep a merge O(n), an algorithm is ran not during merge, but during type registration, that populates a map from non-normalized arrays to normalized types.

This allows the merge to provide a non-normalized array to the type lookup function and still returning a normalized type, while keeping the merge itself O(n).