#68 remove const from ecs_os_api definition
Some compilers place a global const variable in the __TEXT__, const section, which enforces that the variable is readonly. There is no reliable way around that, so for portability, the const is removed.

This makes the API a little bit less safe, but as long as applications don't do anything stupid (change ecs_os_api function pointers halfway the program) things should work fine.