#124 fix bug when adding trait to entity with switch
Switch columns must be stored after trait columns (and regular columns) so that they can be easily addressed with a single offset.

This can be ensured by making the SWITCH role larger than the TRAIT role.

The CASE role needs to be larger than the SWITCH role, so that the SWITCH is placed before the CASE in a type, which ensures that a SWITCH can be initialized with a case in a type.