`flecs::entity_t` vs `flecs::entity`
On my machine, with MSVC, I have to change the parameter types to `flecs::entity_t` here in order for it to compile. I am not certain if the C++ API functor signature should be changed, if this documentation should be updated, or if something is wrong with my setup.