Update flecs::ref to store world instead of stage (#913)
* update ecs_get_ref_id() to handle staged world

Encountered a case using the C++ api where an assert() hit when using
`get_ref().get()`.  This was because the world inside the `flecs::ref`
was actually a stage, and not the world.

Updated `ecs_get_ref_id()` to convert stage to world during lookup.

* Revert "update ecs_get_ref_id() to handle staged world"

This reverts commit 1ca55e46eeb0070ae7df5d16afb517ab04f24226.

* convert stage to world in flecs::ref constructor

From feedback on PR, moved the conversion of stage into workd to the
constructor for flecs::ref.