ecs_get_mut_id() returns live data where possible
From discussion in issue #911, `get_mut()` will now return live data
when the component is present in the entity, or the stage is not
deferred.

When the component does not already exist on the entity, and the system
is deferred, the component will be created in the command queue and
returned.

This commit does not fix the problem of multiple `get_mut()` calls getting
different pointers prior to a pipeline sync.