Include query plan used to populate cache in REST reply
Summary:
This diff adds the plan for the query that is used to populate a query cache to a query_info REST reply. By default the plan of a cached query only shows the instruction that is used to iterate the query cache, but not the instructions that are used to populate the cache, which can make debugging harder.

This change is visible in the explorer query inspector (the "Cache plan" is new):
 {F1907628875}

Differential Revision: D63866489