#1397 Add support for type checked term_at<T> and term_at<T>(i)
First method is simple loop over terms to find term for T. Its slower
than using index directly, but sometimes type safety is more preferred
and when the cost is only on initialization.

Second method is same as term_at(i) but with type checked assert. For
middle ground to still keep the safety but also performance.

This is mostly to mirror the C# binding PR where this is slightly more
useful but maybe cpp would benefit too:

https://github.com/BeanCheeseBurrito/Flecs.NET/pull/52

Signed-off-by: Tomas Slusny <slusnucky@gmail.com>