#1853 Don't create component records unnecessarily
Currently registering any component creates a (Component, *) component record, which is used in the code to keep track of trait flags (bitsets for quickly checking traits). This is however a very heavy weight mechanism just for checking traits, and in most scenarios a component won't be used in a pair.

This PR implements a different mechanism to get trait flags for when a component record doesn't exist yet, which allows the code to not create the (Component, *) records, which saves memory and reduces time spent on component registration.